requirements.txt
tensorflow
Pillow
pandas
numpy
opencv-python